requirements.txt
python-binance
requests
anthropic
websocket-client